feat: jwt 권한 이용한 crud 허용 기능

+++ server/wiselife/src/main/java/be/wiselife/challengetalk/service/ChallengeTalkService.java
@@ -1,5 +1,6 @@
 package be.wiselife.challengetalk.service;
 
+import be.wiselife.challenge.entity.Challenge;
 import be.wiselife.challenge.service.ChallengeService;
 import be.wiselife.challengetalk.entity.ChallengeTalk;
 import be.wiselife.challengetalk.repository.ChallengeTalkRepository;
@@ -24,22 +25,22 @@ public class ChallengeTalkService {
         this.challengeService = challengeService;
     }
 
-    public ChallengeTalk createChallengeTalk(ChallengeTalk challengeTalk, Long challengeId){
-        challengeTalk.setChallenge(challengeService.getChallenge(challengeId));
+    public ChallengeTalk createChallengeTalk(ChallengeTalk challengeTalk, Long challengeId, Member loginMember){
+        Challenge challenge = challengeService.getChallenge(challengeId);
+
+        challengeTalk.setChallenge(challenge);
+        challengeTalk.setMemberId(loginMember.getMemberId());
+        challengeTalk.setCreate_by_member(loginMember.getMemberName());
+
+        challenge.addChallengeTalk(challengeTalk);
 
         return saveChallengeTalk(challengeTalk);
     }
 
-    public ChallengeTalk updateChallengeTalk(ChallengeTalk changedChallengeTalk, String tryingMemberEmail) {
-
+    public ChallengeTalk updateChallengeTalk(ChallengeTalk changedChallengeTalk, Member loginMember) {
+        //사용자 권한 확인
         ChallengeTalk savedChallengeTalk = findChallengeTalkById(changedChallengeTalk.getChallengeTalkId());
-
-        /*유저 권한 확인
-        * 댓글 작성자 email & 수정 요청자 email 비교
-        * */
-        if(!memberService.isVerifiedMember(findChallengeTalkWriterEmail(savedChallengeTalk.getMemberId()), tryingMemberEmail)){
-            throw new BusinessLogicException(ExceptionCode.FORBIDDEN_MEMBER);
-        }
+        checkUserAuthorization(savedChallengeTalk, loginMember);
 
         /*수정 로직*/
         Optional.ofNullable(changedChallengeTalk.getChallengeTalkBody())
@@ -65,8 +66,23 @@ public class ChallengeTalkService {
     }
 
 
-    public void deleteChallengeTalk(Long challengeTalkId) {
+    public void deleteChallengeTalk(Long challengeTalkId, Member loginMember) {
+
         ChallengeTalk challengeTalk = findChallengeTalkById(challengeTalkId);
+        checkUserAuthorization(challengeTalk, loginMember);
+
         challengeTalkRepository.delete(challengeTalk);
+    }
+
+    /**
+     * 챌랜지 댓글 관련 유저의 권한 확인
+     * 챌린지 댓글 수정, 삭제 시도시 사용한다.
+     * @param challengeTalk 변경을 시도하는 챌린지 댓글
+     * @param loginMember 챌랜지 댓글 변경을 시도하는 맴버
+     */
+    private void checkUserAuthorization(ChallengeTalk challengeTalk, Member loginMember){
+        if(!memberService.isVerifiedMember(challengeTalk.getMemberId(), loginMember.getMemberId())){
+            throw new BusinessLogicException(ExceptionCode.FORBIDDEN_MEMBER);
+        }
     }
 }
